backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.0.2
python-dotenv==1.0.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
httpx>=0.24.0,<0.25.0
cryptography==41.0.7
msal==1.24.1
requests==2.31.0
aiohttp==3.9.1
pytz==2023.3
email-validator==2.1.0
icalendar==5.0.11